Reorganize test helpers and gate test-only modules
Split parquet integration_tests into support/parquet helpers plus
co-located unit_tests; extract gen_unique_grpc_port to common/ports
and cfg-gate transport/common test_utils; move runtime/tests benchmarks
to tests/benchmark and stream_task_actor_test next to stream_task_actor.
