+++ src/common/mod.rs
@@ -1,4 +1,6 @@
 pub mod message;
+pub mod ports;
+#[cfg(test)]
 pub mod test_utils;
 pub mod key;
 pub mod types;
