+++ src/runtime/functions/sink/in_memory_storage_sink.rs
@@ -224,7 +224,8 @@ mod tests {
     use arrow::datatypes::{Field, Schema, TimeUnit};
     use arrow::record_batch::RecordBatch;
     use crate::common::message::{BaseMessage, KeyedMessage};
-    use crate::common::test_utils::{create_test_string_batch, gen_unique_grpc_port};
+    use crate::common::ports::gen_unique_grpc_port;
+    use crate::common::test_utils::create_test_string_batch;
     use crate::common::Key;
     use crate::storage::in_memory_storage_grpc_server::InMemoryStorageServer;
