+++ src/tests/benchmark/mod.rs
@@ -1,6 +1,3 @@
-#![cfg(test)]
-
-pub mod stream_task_actor_test;
 pub mod watermark_streaming_benchmark_test;
 pub mod window_operator_benchmark;
 pub mod window_request_operator_benchmark;
