+++ src/tests/mod.rs
@@ -1,5 +1,6 @@
 #![cfg(test)]
 
+pub mod benchmark;
 pub mod docker;
 pub mod inprocess;
 pub mod kube;
